[RFC]: i386/tdx: Fix cpuid[1].EDX.HT=0 when UP
When UP and TDX enabled, cpuid[1].EDX.HT is wrongly set to 1.  This patch
is a workaround patch to unblock guest work to handle the CPUID[1].EDX.HT
bit.  The clean fix to handle any cpuid is needed.

The root cause is that cpuid of the case of configured && native isn't
handled properly.  The functions, tdx_get_supported_cpuid() and
tdx_cap_cpuid_config, unconditionally ignores configured and sets the bit
based on native.

Fixes: 6dc2265a09c2 ("i386/tdx: Adjust the supported CPUID based on TDX restrictions")
Reported-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>
Signed-off-by: Isaku Yamahata <isaku.yamahata@intel.com>